+++ app/src/main/java/com/example/laboratorio_pdm_c2/Dao/CategoriaDao.java
@@ -23,10 +23,10 @@ public interface CategoriaDao {
     Categoria getCategoria(int idcategoria);
 
     @Update
-    long updateCategoria(Categoria categoria);
+    int updateCategoria(Categoria categoria);
 
     @Delete
-    long deleteCategoria(Categoria categoria);
+    int deleteCategoria(Categoria categoria);
 
 
 }
